Flyttet interface i egen fil
